aws/cdk/requirements.txt
aws-cdk-lib
constructs
aws-cdk.core 
aws-cdk.aws-ecs 
aws-cdk.aws-ec2 
aws-cdk.aws-elasticloadbalancingv2 
aws-cdk.aws-autoscaling

